Add: Canvas and Panels for choosing Tutorials. Fix: LOD distance in editor and Quest 2 not being the same.
Added UI interaction for controller.
Added Backend for UI Panel switching.
Added different panels to be shown after completing tutorial.
Made Godrays a prefab.
Fixed the problem, where the distance for different LODs in the editor scene and in the Quest 2 were not the same. (Changed in the Quality Settings)